Removed IntervalTest.scala Moved abbreviated version of it into ElementTest.scala
